+++ api/src/main/java/com/juhai/api/controller/ProjectController.java
@@ -57,10 +57,10 @@ public class ProjectController {
                 JSONObject temp = new JSONObject();
                 temp.put("projectId", project.getId());
                 temp.put("projectName", project.getProjectName());
-                temp.put("projectAmount", project.getProjectAmount());
+                temp.put("projectAmount", project.getProjectAmount().intValue());
                 temp.put("incomeRate", project.getIncomeRate());
                 temp.put("limitTime", project.getLimitTime());
-                temp.put("minAmount", project.getMinAmount());
+                temp.put("minAmount", project.getMinAmount().intValue());
                 temp.put("startTime", project.getStartTime());
                 temp.put("schedule", project.getSchedule());
                 temp.put("guaranteeCompany", map.get("guarantee_company"));
@@ -93,10 +93,10 @@ public class ProjectController {
         JSONObject temp = new JSONObject();
         temp.put("projectId", project.getId());
         temp.put("projectName", project.getProjectName());
-        temp.put("projectAmount", project.getProjectAmount());
+        temp.put("projectAmount", project.getProjectAmount().intValue());
         temp.put("incomeRate", project.getIncomeRate());
         temp.put("limitTime", project.getLimitTime());
-        temp.put("minAmount", project.getMinAmount());
+        temp.put("minAmount", project.getMinAmount().intValue());
         temp.put("startTime", project.getStartTime());
         temp.put("schedule", project.getSchedule());
         temp.put("guaranteeCompany", map.get("guarantee_company"));
