+++ api/src/main/java/com/juhai/api/controller/UserController.java
@@ -87,7 +87,7 @@ public class UserController {
         JSONObject temp = new JSONObject();
         temp.put("userId", user.getId());
         temp.put("userName", user.getUserName());
-        temp.put("balance", user.getBalance());
+        temp.put("balance", user.getBalance().intValue());
         temp.put("realName", user.getRealName());
         temp.put("idCard", DesensitizedUtil.idCardNum(user.getIdCard(), 2, 2));
         temp.put("inviteCode", user.getInviteCode());
@@ -121,8 +121,8 @@ public class UserController {
             waitReturnPrincipal = NumberUtil.add(waitReturnPrincipal, order.getAmount());
         }
         
-        temp.put("waitReturnInterest", waitReturnInterest);
-        temp.put("waitReturnPrincipal", waitReturnPrincipal);
+        temp.put("waitReturnInterest", waitReturnInterest.intValue());
+        temp.put("waitReturnPrincipal", waitReturnPrincipal.intValue());
 
         return R.ok().put("data", temp);
     }
@@ -148,8 +148,8 @@ public class UserController {
             waitReturnPrincipal = NumberUtil.add(waitReturnPrincipal, order.getAmount());
         }
 
-        temp.put("waitReturnInterest", waitReturnInterest);
-        temp.put("waitReturnPrincipal", waitReturnPrincipal);
+        temp.put("waitReturnInterest", waitReturnInterest.intValue());
+        temp.put("waitReturnPrincipal", waitReturnPrincipal.intValue());
 
         return R.ok().put("data", temp);
     }
@@ -369,7 +369,7 @@ public class UserController {
                 JSONObject obj = new JSONObject();
                 obj.put("remark", temp.getRemark());
 //                obj.put("remark", map.getOrDefault(temp.getOptType(), "-"));
-                obj.put("amount", temp.getOptAmount());
+                obj.put("amount", temp.getOptAmount().intValue());
                 obj.put("optTime", temp.getOptTime());
                 arr.add(obj);
             }
@@ -396,7 +396,7 @@ public class UserController {
                 JSONObject obj = new JSONObject();
                 obj.put("projectName", temp.getProjectName());
                 obj.put("orderNo", temp.getOrderNo());
-                obj.put("amount", temp.getAmount());
+                obj.put("amount", temp.getAmount().intValue());
                 obj.put("status", temp.getStatus());
                 arr.add(obj);
             }
@@ -426,7 +426,7 @@ public class UserController {
                 obj.put("projectName", temp.getProjectName());
                 obj.put("returnTime", temp.getActualReturnTime());
                 obj.put("status", temp.getStatus());
-                obj.put("amount", NumberUtil.add(temp.getAmount(), temp.getActualReturnAmount()));
+                obj.put("amount", NumberUtil.add(temp.getAmount(), temp.getActualReturnAmount()).intValue());
                 arr.add(obj);
             }
             page.setList(arr);
@@ -453,7 +453,7 @@ public class UserController {
                 obj.put("typeStr", MsgUtil.get("system.user.deposittype"));
                 obj.put("time", temp.getOptTime());
                 obj.put("status", temp.getStatus());
-                obj.put("amount", temp.getOptAmount());
+                obj.put("amount", temp.getOptAmount().intValue());
                 arr.add(obj);
             }
             page.setList(arr);
@@ -479,7 +479,7 @@ public class UserController {
                 JSONObject obj = new JSONObject();
                 obj.put("time", temp.getOptTime());
                 obj.put("status", temp.getStatus());
-                obj.put("amount", temp.getOptAmount());
+                obj.put("amount", temp.getOptAmount().intValue());
                 arr.add(obj);
             }
             page.setList(arr);
